More places added - some formatting of text

+++ TopPlacesToVisitinScotland/app/src/test/java/com/codeclan/topplacestovisitinscotland/PlacesTest.java
@@ -18,7 +18,8 @@ public class PlacesTest {
 
     @Test
     public void getRankingTest() {
-        assertEquals(7, places.getRanking() );
+
+        assertEquals((Integer)7, places.getRanking());
     }
 
     @Test
